Added forkIO to chanToHandle patch from Ganesh Sittampalam.
He wrote:

For some calling contexts, chanToHandle needs to fork and then close
the handle when done. This behaviour needs to be independent of what
kind of Channel we are writing out.

Hi John,

I've been playing with HSH, in an attempt to replace a simple shell
script with a Haskell program, and I've run into a few issues.

I needed to add forkIO to chanToHandle (ChanBSL ...). This makes sense
to me, as BSL.hPut h s will dump the entirety of s into h immediately,
and if h is the input to a pipe then it would block after the buffer
fills up until some output is consumed. Without using a separate
thread, nothing can consume the output so we have a deadlock. I think
the same change should be made to the ChanString case too.

In fact there are some calls to chanToHandle where it is not
appropriate to fork even in the Handle -> Handle case (which it
already did); for example when it is called with stdout as the target
handle. I therefore added a boolean flag to indicate whether to fork
or not - see the attached patch.

Next, (after making those changes) my Haskell program needs to be
compiled with -threaded to not deadlock (though it works ok in
ghci/runhaskell, possibly because those are themselves built
-threaded). After some digging, I discovered a couple of different
reasons for this.

(a) When chanToHandle forks, it also needs to hClose the handle,
otherwise it ends up blocked indefinitely which means the command it
was the input pipe of never terminates. In the threaded runtime, it
ends up being GCed and closed by a finalizer, but that never happens
in the non-threaded runtime.

(b) There's a GHC bug introduced in 6.8.3 which means that
non-blocking IO doesn't actually work properly in the non-threaded
runtime. I've reported this.

It's worth making things work properly with the non-threaded runtime,
btw, because it seems to perform significantly better - the threaded
runtime seems to have quite a bit more overhead, possibly related to
use of futexes, on each fork/exec.

I also found the strictness/laziness of the various ways of calling
'run' quite unintuitive. If it gives me a lazy bytestring, I expect
that to be really lazy! It took me some time to realise that I had to
use the IO (ByteString, IO (...)) version. Personally, I feel that the
overloading of run is rather messy; I think it would be better to
split it into strict and lazy variants, where the lazy variants always
give you the IO (String, ExitCode) return element and the strict
variants never do.